refactor : Go to Profile
